Split src/types into current and legacy, update platform icon types
