+++ src/components/gamma/stack/header/StackHeaderConfig.ios.types.ts
@@ -1,5 +1,5 @@
 import type { ReactElement } from 'react';
-import type { PlatformIconIOS } from '../../../../types';
+import type { PlatformIconIOS } from '../../../shared/types';
 import type { StackHeaderMenuIOS } from './ios/StackHeaderMenu.ios.types';
 
 /**
