+++ src/legacy/components/Screen.tsx
@@ -5,7 +5,7 @@ import { Animated, View, Platform } from 'react-native';
 
 import TransitionProgressContext from '../TransitionProgressContext';
 import DelayedFreeze from './helpers/DelayedFreeze';
-import { ScreenProps } from '../../types';
+import { ScreenProps } from '../types';
 
 import {
   freezeEnabled,
