+++ src/legacy/components/ScreenStack.tsx
@@ -7,7 +7,7 @@ import {
   GestureProviderProps,
   GoBackGesture,
   ScreenStackProps,
-} from '../../types';
+} from '../types';
 import { GHContext, RNSScreensRefContext } from '../contexts';
 import warnOnce from 'warn-once';
