+++ src/legacy/components/helpers/prepareHeaderBarButtonItems.ts
@@ -1,8 +1,5 @@
 import { Image, processColor } from 'react-native';
-import {
-  HeaderBarButtonItem,
-  HeaderBarButtonItemWithMenu,
-} from 'react-native-screens/types';
+import { HeaderBarButtonItem, HeaderBarButtonItemWithMenu } from '../../types';
 
 const prepareMenu = (
   menu: HeaderBarButtonItemWithMenu['menu'],
