+++ src/legacy/contexts.tsx
@@ -1,5 +1,5 @@
 import React, { PropsWithChildren } from 'react';
-import { GestureProviderProps, ScreensRefsHolder } from '../types';
+import { GestureProviderProps, ScreensRefsHolder } from './types';
 
 export const GHContext = React.createContext(
   (props: PropsWithChildren<GestureProviderProps>) => <>{props.children}</>,
