+++ src/legacy/gesture-handler/ScreenGestureDetector.tsx
@@ -22,7 +22,7 @@ import {
   checkIfTransitionCancelled,
   getAnimationForTransition,
 } from './constraints';
-import { GestureProviderProps } from '../../types';
+import { GestureProviderProps } from '../types';
 
 // The detector is disabled to work around issue with pressables
 // losing focus. See https://github.com/software-mansion/react-native-screens/pull/2819
