+++ src/legacy/reanimated/ReanimatedScreen.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { InnerScreen } from '../components/Screen';
-import { ScreenProps } from '../../types';
+import { ScreenProps } from '../types';
 
 // @ts-ignore file to be used only if `react-native-reanimated` available in the project
 import Animated from 'react-native-reanimated';
