+++ src/legacy/reanimated/ReanimatedScreenProvider.tsx
@@ -1,7 +1,7 @@
 import React, { PropsWithChildren } from 'react';
 import { View } from 'react-native';
 import { ScreenContext } from '../components/Screen';
-import { ScreenProps } from '../../types';
+import { ScreenProps } from '../types';
 import ReanimatedNativeStackScreen from './ReanimatedNativeStackScreen';
 import AnimatedScreen from './ReanimatedScreen';
