+++ src/types.ts
@@ -1,0 +1,15 @@
+export type {
+  BlurEffect,
+  ColorScheme,
+  Direction,
+  InterfaceOrientation,
+  ScrollEdgeEffect,
+  UserInterfaceStyle,
+  PlatformIconShared,
+  PlatformIconAndroidDrawableResource,
+  PlatformIconIOSTemplate,
+  PlatformIconIOSSfSymbol,
+  PlatformIconIOSXcasset,
+  PlatformIconIOS,
+  PlatformIconAndroid,
+} from './components/shared/types';
